add money-rails to handle money, add fields to expenses/show

+++ db/migrate/20140209223434_create_expenses.rb
@@ -7,7 +7,7 @@ class CreateExpenses < ActiveRecord::Migration
       t.string :note
       t.date :date
       t.boolean :avoidable
-      t.integer :cost
+      t.money :cost
 
       t.timestamps
     end
